Set up killfloor for players and enemies
Destroys enemies outright, and the player takes damage and is respawned to the platform

+++ Project/Robot Rumble/Assets/Scripts/Health.cs
@@ -20,7 +20,7 @@ public class Health : MonoBehaviour
         Destroy(gameObject);
     }
 
-    public float GetHealth(){
+    public int GetHealth(){
         return health;
     }
 }
